对删除面试官进行校验，无法删除抛出自定义异常

+++ service/src/main/java/com/zql/service/impl/InterviewerServiceImpl.java
@@ -3,9 +3,13 @@ package com.zql.service.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.mysql.cj.util.StringUtils;
 import com.zql.bo.InterviewerBo;
+import com.zql.exception.GraceException;
+import com.zql.grace.result.ResponseStatusEnum;
 import com.zql.mapper.InterviewerMapper;
 import com.zql.pojo.Interviewer;
 import com.zql.service.InterviewerService;
+import com.zql.service.JobService;
+import com.zql.service.QuestionLibService;
 import jakarta.annotation.Resource;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
@@ -18,6 +22,12 @@ public class InterviewerServiceImpl implements InterviewerService {
 
     @Resource
     private InterviewerMapper interviewerMapper;
+
+    @Resource
+    private JobService jobService;
+
+    @Resource
+    private QuestionLibService questionLibService;
 
     @Override
     public void CreateOrUpdate(InterviewerBo interviewerBo){
@@ -43,8 +53,15 @@ public class InterviewerServiceImpl implements InterviewerService {
     @Override
     public void delete(String id) {
 
-        //todo  检验面试官是否关联其他的工作职位
+        boolean jobFlag = jobService.isJObContainsInterviewer(id);
+        boolean questionFlag = questionLibService.isQuestionContainsInterviewer(id);
 
-        interviewerMapper.deleteById(id);
+        if(jobFlag || questionFlag){
+            GraceException.display(ResponseStatusEnum.CAN_NOT_DELETE_INTERVIEWER);
+        }
+        else{
+            interviewerMapper.deleteById(id);
+        }
+
     }
 }
